espresso_types

Module v0

Re-exports

pub use v0_99::AccountQueryData;

pub use v0_99::BlockMerkleCommitment;

pub use v0_99::BlockMerkleTree;

pub use v0_99::BuilderSignature;

pub use v0_99::ChainId;

pub use v0_99::Delta;

pub use v0_99::FeeAccount;

pub use v0_99::FeeAccountProof;

pub use v0_99::FeeAmount;

pub use v0_99::FeeInfo;

pub use v0_99::FeeMerkleCommitment;

pub use v0_99::FeeMerkleProof;

pub use v0_99::FeeMerkleTree;

pub use v0_99::Index;

pub use v0_99::Iter;

pub use v0_99::L1BlockInfo;

pub use v0_99::L1Client;

pub use v0_99::L1ClientOptions;

pub use v0_99::L1Snapshot;

pub use v0_99::NamespaceId;

pub use v0_99::NsIndex;

pub use v0_99::NsIter;

pub use v0_99::NsPayload;

pub use v0_99::NsPayloadBuilder;

pub use v0_99::NsPayloadByteLen;

pub use v0_99::NsPayloadOwned;

pub use v0_99::NsPayloadRange;

pub use v0_99::NsTable;

pub use v0_99::NsTableBuilder;

pub use v0_99::NsTableValidationError;

pub use v0_99::NumNss;

pub use v0_99::NumTxs;

pub use v0_99::NumTxsRange;

pub use v0_99::NumTxsUnchecked;

pub use v0_99::Payload;

pub use v0_99::PayloadByteLen;

pub use v0_99::Transaction;

pub use v0_99::TxIndex;

pub use v0_99::TxIter;

pub use v0_99::TxPayload;

pub use v0_99::TxPayloadRange;

pub use v0_99::TxProof;

pub use v0_99::TxTableEntries;

pub use v0_99::TxTableEntriesRange;

pub use v0_99::Upgrade;

pub use v0_99::UpgradeType;

pub use v0_99::UpgradeMode;

pub use v0_99::TimeBasedUpgrade;

pub use v0_99::ViewBasedUpgrade;

pub use v0_99::BlockSize;

pub use crate::v0_1::BLOCK_MERKLE_TREE_HEIGHT;

pub use crate::v0_1::FEE_MERKLE_TREE_HEIGHT;

pub use crate::v0_1::NS_ID_BYTE_LEN;

pub use crate::v0_1::NS_OFFSET_BYTE_LEN;

pub use crate::v0_1::NUM_NSS_BYTE_LEN;

pub use crate::v0_1::NUM_TXS_BYTE_LEN;

pub use crate::v0_1::TX_OFFSET_BYTE_LEN;

Modules

config

header 🔒

impls 🔒

mock

nsproof 🔒

traits

This module contains all the traits used for building the sequencer types. It also includes some trait implementations that cannot be implemented in an external crate.

utils 🔒

v0_1

v0_2

v0_3

v0_99

Macros

assert_eq_all_versions_of_type 🔒

define_modules 🔒

reexport_latest_version_of_type 🔒

reexport_unchanged_types 🔒

Re-export types which have not changed across any minor version.

with_minor_versions 🔒

Structs

BackoffParams

EpochCommittees

Type to describe DA and Stake memberships

GenesisHeader

Information about the genesis state which feeds into the genesis block header.

NodeState

Represents the immutable state of a node.

ParseDurationError

ParseSizeError

Ratio

SeqTypes

SequencerVersions

SolverAuctionResultsProvider

Auction Results provider holding the Url of the solver in order to fetch auction results.

Timestamp

ValidatedState

State to be validated by replicas.

Enums

BuilderValidationError

Possible builder validation failures

FeeError

Possible charge fee failures

Header

Each variant represents a specific minor version header.

NsProof

Each variant represents a specific version of a namespace proof.

ParseRatioError

ProposalValidationError

Possible proposal validation failures

StateValidationError

This enum is not used in code but functions as an index of possible validation errors.

Update

Constants

BACKOFF_FACTOR

BACKOFF_JITTER

MAX_RETRY_DELAY

MIN_RETRY_DELAY

Functions

get_l1_deposits

parse_duration

parse_size

retain_accounts

Get a partial snapshot of the given fee state, which contains only the specified accounts.

upgrade_commitment_map

Type Aliases

EpochVersion

Event

FeeVersion

Leaf

Leaf2

MarketplaceVersion

MockSequencerVersions

NetworkConfig

PrivKey

PubKey

V0_0

V0_1