Function espresso_types::v0::retain_accounts

pub fn retain_accounts(
    state: &FeeMerkleTree,
    accounts: impl IntoIterator<Item = FeeAccount>,
) -> Result<FeeMerkleTree>

Get a partial snapshot of the given fee state, which contains only the specified accounts.

Fails if one of the requested accounts is not represented in the original state.