espresso_types::v0::impls::block::namespace_payload::ns_payload
Module ns_payload_owned

Crazy boilerplate code to make it so that [NsPayloadOwned] is to NsPayload as Vec<T> is to [T]. See How can I create newtypes for an unsized type and its owned counterpart (like str and String) in safe Rust? - Stack Overflow