espresso_types::v0::impls::stake_table
Function from_l1_events

pub fn from_l1_events<I: Iterator<Item = StakeTableEvent>>(
    events: I,
) -> Result<IndexMap<Address, Validator<BLSPubKey>>>

Create the consensus and DA stake tables from L1 events

This is a pure function, to make it easily testable.

We expect have at most a few hundred EVM events in the PermissionedStakeTable contract over the liftetime of the contract so it should not significantly affect performance to fetch all events and perform the computation in this functions once per epoch.