espresso_types::v0::reward_mt

Type Alias RewardNonMembershipProof

pub type RewardNonMembershipProof = RewardMerkleProof;

Non-membership proof for an account not in the tree.

Proves that a specific account does NOT exist in the tree. In a universal Merkle tree, this is structurally identical to a membership proof, showing the path to where the account would be if it existed (an empty position).

Aliased Type

pub struct RewardNonMembershipProof {
    pub pos: RewardAccountV2,
    pub proof: Vec<MerkleNode<RewardAmount, RewardAccountV2, KeccakNode>>,
}

Fields
pos: RewardAccountV2

Proof of inclusion for element at index pos

proof: Vec<MerkleNode<RewardAmount, RewardAccountV2, KeccakNode>>

Nodes of proof path, from root to leaf