espresso_types::v0
Module traits

This module contains all the traits used for building the sequencer types. It also includes some trait implementations that cannot be implemented in an external crate.

Structs

NullEventConsumer

Traits

EventConsumer

FromNsPayloadBytes

Data that can be deserialized from a subslice of namespace payload bytes.

FromStringOrInteger

Types which can be deserialized from either integers or strings.

NsPayloadBytesRange

Specifies a subslice of namespace payload bytes to read.

PersistenceOptions

SequencerPersistence

StateCatchup