Trait espresso_types::v0::traits::NsPayloadBytesRange

pub trait NsPayloadBytesRange<'a> {
    type Output: FromNsPayloadBytes<'a>;

    // Required method
    fn ns_payload_range(&self) -> Range<usize>;
}

Specifies a subslice of namespace payload bytes to read.

Companion trait for FromNsPayloadBytes, which holds data that can be deserialized from that subslice of bytes.

Required Associated Types

type Output: FromNsPayloadBytes<'a>

Required Methods

fn ns_payload_range(&self) -> Range<usize>

Range relative to this ns payload