espresso_types::v0::traits

Trait StateCatchup

pub trait StateCatchup: Send + Sync {
    // Required methods
    fn try_fetch_leaves<'life0, 'async_trait>(
        &'life0 self,
        retry: usize,
        height: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Leaf2>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn try_fetch_accounts<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        retry: usize,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        fee_merkle_tree_root: FeeMerkleCommitment,
        account: &'life2 [FeeAccount],
    ) -> Pin<Box<dyn Future<Output = Result<FeeMerkleTree>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn try_remember_blocks_merkle_tree<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        retry: usize,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        mt: &'life2 mut BlockMerkleTree,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn try_fetch_chain_config<'life0, 'async_trait>(
        &'life0 self,
        retry: usize,
        commitment: Commitment<ChainConfig>,
    ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn try_fetch_reward_accounts<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        retry: usize,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        reward_merkle_tree_root: RewardMerkleCommitment,
        account: &'life2 [RewardAccount],
    ) -> Pin<Box<dyn Future<Output = Result<RewardMerkleTree>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn backoff(&self) -> &BackoffParams;
    fn name(&self) -> String;

    // Provided methods
    fn fetch_leaf<'life0, 'async_trait>(
        &'life0 self,
        height: u64,
        stake_table: Vec<PeerConfig<SeqTypes>>,
        success_threshold: U256,
    ) -> Pin<Box<dyn Future<Output = Result<Leaf2>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn fetch_accounts<'life0, 'life1, 'async_trait>(
        &'life0 self,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        fee_merkle_tree_root: FeeMerkleCommitment,
        accounts: Vec<FeeAccount>,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<FeeAccountProof>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn remember_blocks_merkle_tree<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        mt: &'life2 mut BlockMerkleTree,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn fetch_chain_config<'life0, 'async_trait>(
        &'life0 self,
        commitment: Commitment<ChainConfig>,
    ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn fetch_reward_accounts<'life0, 'life1, 'async_trait>(
        &'life0 self,
        instance: &'life1 NodeState,
        height: u64,
        view: ViewNumber,
        reward_merkle_tree_root: RewardMerkleCommitment,
        accounts: Vec<RewardAccount>,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<RewardAccountProof>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Required Methods

fn try_fetch_leaves<'life0, 'async_trait>( &'life0 self, retry: usize, height: u64, ) -> Pin<Box<dyn Future<Output = Result<Vec<Leaf2>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn try_fetch_accounts<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, retry: usize, instance: &'life1 NodeState, height: u64, view: ViewNumber, fee_merkle_tree_root: FeeMerkleCommitment, account: &'life2 [FeeAccount], ) -> Pin<Box<dyn Future<Output = Result<FeeMerkleTree>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Try to fetch the given accounts state, failing without retrying if unable.

fn try_remember_blocks_merkle_tree<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, retry: usize, instance: &'life1 NodeState, height: u64, view: ViewNumber, mt: &'life2 mut BlockMerkleTree, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Try to fetch and remember the blocks frontier, failing without retrying if unable.

fn try_fetch_chain_config<'life0, 'async_trait>( &'life0 self, retry: usize, commitment: Commitment<ChainConfig>, ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn try_fetch_reward_accounts<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, retry: usize, instance: &'life1 NodeState, height: u64, view: ViewNumber, reward_merkle_tree_root: RewardMerkleCommitment, account: &'life2 [RewardAccount], ) -> Pin<Box<dyn Future<Output = Result<RewardMerkleTree>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Try to fetch the given accounts state, failing without retrying if unable.

fn backoff(&self) -> &BackoffParams

fn name(&self) -> String

Provided Methods

fn fetch_leaf<'life0, 'async_trait>( &'life0 self, height: u64, stake_table: Vec<PeerConfig<SeqTypes>>, success_threshold: U256, ) -> Pin<Box<dyn Future<Output = Result<Leaf2>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn fetch_accounts<'life0, 'life1, 'async_trait>( &'life0 self, instance: &'life1 NodeState, height: u64, view: ViewNumber, fee_merkle_tree_root: FeeMerkleCommitment, accounts: Vec<FeeAccount>, ) -> Pin<Box<dyn Future<Output = Result<Vec<FeeAccountProof>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetch the given list of accounts, retrying on transient errors.

fn remember_blocks_merkle_tree<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, instance: &'life1 NodeState, height: u64, view: ViewNumber, mt: &'life2 mut BlockMerkleTree, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Fetch and remember the blocks frontier, retrying on transient errors.

fn fetch_chain_config<'life0, 'async_trait>( &'life0 self, commitment: Commitment<ChainConfig>, ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn fetch_reward_accounts<'life0, 'life1, 'async_trait>( &'life0 self, instance: &'life1 NodeState, height: u64, view: ViewNumber, reward_merkle_tree_root: RewardMerkleCommitment, accounts: Vec<RewardAccount>, ) -> Pin<Box<dyn Future<Output = Result<Vec<RewardAccountProof>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetch the given list of accounts, retrying on transient errors.