espresso_types::v0
Type Alias Leaf2

pub type Leaf2 = Leaf2<SeqTypes>;

Aliased Type

struct Leaf2 {
    pub view_change_evidence: Option<ViewChangeEvidence2<SeqTypes>>,
    pub next_drb_result: Option<[u8; 32]>,
    pub with_epoch: bool,
    /* private fields */
}

Fields
view_change_evidence: Option<ViewChangeEvidence2<SeqTypes>>

Possible timeout or view sync certificate. If the justify_qc is not for a proposal in the immediately preceding view, then either a timeout or view sync certificate must be attached.

next_drb_result: Option<[u8; 32]>

The DRB result for the next epoch.

This is required only for the last block of the epoch. Nodes will verify that it’s consistent with the result from their computations.

with_epoch: bool

Indicates whether or not epochs were enabled.