Type Alias espresso_types::v0::NetworkConfig

pub type NetworkConfig = NetworkConfig<PubKey>;

Aliased Type

struct NetworkConfig {

    pub rounds: usize,
    pub indexed_da: bool,
    pub transactions_per_round: usize,
    pub manual_start_password: Option<String>,
    pub num_bootrap: usize,
    pub next_view_timeout: u64,
    pub view_sync_timeout: Duration,
    pub builder_timeout: Duration,
    pub data_request_delay: Duration,
    pub node_index: u64,
    pub seed: [u8; 32],
    pub transaction_size: usize,
    pub start_delay_seconds: u64,
    pub key_type_name: String,
    pub libp2p_config: Option<Libp2pConfig>,
    pub config: HotShotConfig<VerKey>,
    pub cdn_marshal_address: Option<String>,
    pub combined_network_config: Option<CombinedNetworkConfig>,
    pub commit_sha: String,
    pub builder: BuilderType,
    pub random_builder: Option<RandomBuilderConfig>,

}

Fields
rounds: usize

number of views to run

indexed_da: bool

whether DA membership is determined by index. if true, the first k nodes to register form the DA committee if false, DA membership is requested by the nodes

transactions_per_round: usize

number of transactions per view

manual_start_password: Option<String>

password to have the orchestrator start the network, regardless of the number of nodes connected.

num_bootrap: usize

number of bootstrap nodes

next_view_timeout: u64

timeout before starting the next view

view_sync_timeout: Duration

timeout before starting next view sync round

builder_timeout: Duration

The maximum amount of time a leader can wait to get a block from a builder

data_request_delay: Duration

time to wait until we request data associated with a proposal

node_index: u64

global index of node (for testing purposes a uid)

seed: [u8; 32]

unique seed (for randomness? TODO)

transaction_size: usize

size of transactions

start_delay_seconds: u64

delay before beginning consensus

key_type_name: String

name of the key type (for debugging)

libp2p_config: Option<Libp2pConfig>

the libp2p config

config: HotShotConfig<VerKey>

the hotshot config

cdn_marshal_address: Option<String>

The address for the Push CDN’s “marshal”, A.K.A. load balancer

combined_network_config: Option<CombinedNetworkConfig>

combined network config

commit_sha: String

the commit this run is based on

builder: BuilderType

builder to use

random_builder: Option<RandomBuilderConfig>

random builder config