Struct espresso_types::v0::v0_1::block::Payload

pub struct Payload {
    pub(crate) raw_payload: Vec<u8>,
    pub(crate) ns_table: NsTable,
}

Raw payload data for an entire block.

A block consists of two sequences of arbitrary bytes:

ns_table: namespace table

ns_payloads: namespace payloads

Any sequence of bytes is a valid ns_table. Any sequence of bytes is a valid ns_payloads. The contents of ns_table determine how to interpret ns_payload.

Namespace table

See NsTable for the format of a namespace table.

Namespace payloads

A concatenation of payload bytes for multiple individual namespaces. Namespace boundaries are dictated by ns_table. See NsPayload for the format of a namespace payload.

Fields
raw_payload: Vec<u8>ns_table: NsTable
Implementations

impl Payload

pub fn ns_table(&self) -> &NsTable

pub fn transaction(&self, index: &Index) -> Option<Transaction>

Like [QueryablePayload::transaction_with_proof] except without the proof.

pub(crate) fn read_ns_payload(&self, range: &NsPayloadRange) -> &NsPayload

pub(crate) fn ns_payload(&self, index: &NsIndex) -> &NsPayload

Convenience wrapper for Self::read_ns_payload.

index is not checked. Use self.ns_table().in_bounds() as needed.

pub(crate) fn byte_len(&self) -> PayloadByteLen

fn from_transactions_sync( transactions: impl IntoIterator<Item = <Self as BlockPayload<SeqTypes>>::Transaction> + Send, chain_config: ChainConfig, _instance_state: &<Self as BlockPayload<SeqTypes>>::Instance, ) -> Result<(Self, <Self as BlockPayload<SeqTypes>>::Metadata), <Self as BlockPayload<SeqTypes>>::Error>

Need a sync version of [BlockPayload::from_transactions] in order to impl [BlockPayload::empty].

impl Payload

pub fn ns_table_mut(&mut self) -> &mut NsTable