espresso_types::v0::v0_1

Struct L1Client

pub struct L1Client {
    pub provider: L1Provider,
    pub transport: SwitchingTransport,
    pub(crate) state: Arc<Mutex<L1State>>,
    pub(crate) sender: Sender<L1Event>,
    pub(crate) receiver: InactiveReceiver<L1Event>,
    pub(crate) update_task: Arc<L1UpdateTask>,
}

An Ethereum provider and configuration to interact with the L1.

This client runs asynchronously, updating an in-memory snapshot of the relevant L1 information each time a new L1 block is published. The main advantage of this is that we can update the L1 state at the pace of the L1, instead of the much faster pace of HotShot consensus.This makes it easy to use a subscription instead of polling for new blocks, vastly reducing the number of L1 RPC calls we make.

Fields
provider: L1Provider

The alloy provider used for L1 communication with wallet and default fillers

transport: SwitchingTransport

Actual transport used in self.provider i.e. the t variable in ProviderBuilder::new().on_client(RpcClient::new(t, is_local))

state: Arc<Mutex<L1State>>

Shared state updated by an asynchronous task which polls the L1.

sender: Sender<L1Event>

Channel used by the async update task to send events to clients.

receiver: InactiveReceiver<L1Event>

Receiver for events from the async update task.

update_task: Arc<L1UpdateTask>

Async task which updates the shared state.

Implementations

impl L1Client

fn with_transport(transport: SwitchingTransport) -> Self

pub fn new(url: Vec<Url>) -> Result<Self>

Construct a new L1 client with the default options.

pub fn anvil(anvil: &AnvilInstance) -> Result<Self>

test only

pub async fn spawn_tasks(&self)

Start the background tasks which keep the L1 client up to date.

pub async fn shut_down_tasks(&self)

Shut down background tasks associated with this L1 client.

The L1 client will still be usable, but will stop updating until start is called again.

fn update_loop(&self) -> impl Future<Output = ()> + use<>

pub async fn snapshot(&self) -> L1Snapshot

Get a snapshot from the l1.

pub async fn wait_for_block(&self, number: u64)

Wait until the highest L1 block number reaches at least number.

This function does not return any information about the block, since the block is not necessarily finalized when it returns. It is only used to guarantee that some block at height number exists, possibly in the unsafe part of the L1 chain.

pub async fn wait_for_finalized_block(&self, number: u64) -> L1BlockInfo

Get information about the given block.

If the desired block number is not finalized yet, this function will block until it becomes finalized.

pub async fn wait_for_finalized_block_with_timestamp( &self, timestamp: U256, ) -> L1BlockInfo

Get information about the first finalized block with timestamp greater than or equal timestamp.

async fn fetch_finalized_block_by_number<'a>( &'a self, state: MutexGuard<'a, L1State>, number: u64, ) -> (MutexGuard<'a, L1State>, L1BlockInfo)

async fn load_and_cache_finalized_block<'a>( &'a self, state: MutexGuard<'a, L1State>, id: BlockId, ) -> (MutexGuard<'a, L1State>, L1BlockInfoWithParent)

pub async fn get_finalized_deposits( &self, fee_contract_address: Address, prev_finalized: Option<u64>, new_finalized: u64, ) -> Vec<FeeInfo>

Get fee info for each Deposit occurring between prev and new. Returns Vec<FeeInfo>

pub async fn is_proxy_contract(&self, proxy_address: Address) -> Result<bool>

Check if the given address is a proxy contract.

pub async fn retry_on_all_providers<Fut>( &self, op: impl Fn() -> Fut, ) -> Result<Fut::Ok, Fut::Error>
where Fut: TryFuture,

pub(crate) fn options(&self) -> &L1ClientOptions

fn metrics(&self) -> &L1ClientMetrics

async fn retry_delay(&self)