espresso_types::v0::v0_1
Struct NsPayload

pub struct NsPayload(pub(crate) [u8]);

Raw binary data for a single namespace’s payload.

Any sequence of bytes is a valid NsPayload.

See module-level documentation types for a full specification of the binary format of a namespace.

Tuple Fields
0: [u8]
Implementations

impl NsPayload

pub(super) fn new_private(p: &[u8]) -> &NsPayload

impl NsPayload

pub fn from_bytes_slice(bytes: &[u8]) -> &NsPayload

pub fn as_bytes_slice(&self) -> &[u8]

pub fn byte_len(&self) -> NsPayloadByteLen

pub fn read<'a, R>(&'a self, range: &R) -> R::Output
where R: NsPayloadBytesRange<'a>,

Read and parse bytes from the ns payload.

Arg range: &R is convertible into a Range<usize> via NsPayloadBytesRange. The payload bytes are parsed into a R::Output via FromNsPayloadBytes.

pub fn iter(&self) -> TxIter

Iterator over all transactions in this namespace.

pub fn export_all_txs(&self, ns_id: &NamespaceId) -> Vec<Transaction>

Return all transactions in this namespace. The namespace ID for each returned Transaction is set to ns_id.

pub fn export_tx( &self, ns_id: &NamespaceId, index: &TxIndex, ) -> Option<Transaction>

Return a transaction from this namespace. Set its namespace ID to ns_id.

Return None if index is out of bounds.

fn read_num_txs(&self) -> NumTxsUnchecked

Private helper. (Could be pub if desired.)

fn iter_from_num_txs(&self, num_txs: &NumTxsUnchecked) -> TxIter

Private helper

fn tx_from_num_txs( &self, ns_id: &NamespaceId, index: &TxIndex, num_txs_unchecked: &NumTxsUnchecked, ) -> Transaction

Private helper