espresso_types::v0::v0_4

Module state

Re-exports

pub use crate::v0::reward_mt::REWARD_MERKLE_TREE_V2_ARITY;

pub use crate::v0::reward_mt::REWARD_MERKLE_TREE_V2_HEIGHT;

pub use crate::v0::reward_mt::RewardMerkleCommitmentV2;

pub use crate::v0::reward_mt::RewardMerkleTreeV2;

Structs

Delta

PermittedRewardMerkleTreeV2

RewardAccountProofV2

A proof of the balance of an account in the fee ledger.

RewardAccountQueryDataV2

RewardAccountV2

Enums

RewardClaimError

RewardMerkleProofV2

Statics

REWARD_MERKLE_TREE_V2_MEMORY_LOCK 🔒

Functions

forgotten_accounts_include

Return true if any of the given accounts have been forgotten in the ValidatedState reward_merkle_tree_v2