espresso_types::v0
Module v0_99

Re-exports

pub use super::v0_1::AccountQueryData;

pub use super::v0_1::BlockMerkleCommitment;

pub use super::v0_1::BlockMerkleTree;

pub use super::v0_1::BlockSize;

pub use super::v0_1::BuilderSignature;

pub use super::v0_1::ChainId;

pub use super::v0_1::Delta;

pub use super::v0_1::FeeAccount;

pub use super::v0_1::FeeAccountProof;

pub use super::v0_1::FeeAmount;

pub use super::v0_1::FeeInfo;

pub use super::v0_1::FeeMerkleCommitment;

pub use super::v0_1::FeeMerkleProof;

pub use super::v0_1::FeeMerkleTree;

pub use super::v0_1::Index;

pub use super::v0_1::Iter;

pub use super::v0_1::L1BlockInfo;

pub use super::v0_1::L1Client;

pub use super::v0_1::L1ClientOptions;

pub use super::v0_1::L1Snapshot;

pub use super::v0_1::NamespaceId;

pub use super::v0_1::NsIndex;

pub use super::v0_1::NsIter;

pub use super::v0_1::NsPayload;

pub use super::v0_1::NsPayloadBuilder;

pub use super::v0_1::NsPayloadByteLen;

pub use super::v0_1::NsPayloadOwned;

pub use super::v0_1::NsPayloadRange;

pub use super::v0_1::NsProof;

pub use super::v0_1::NsTable;

pub use super::v0_1::NsTableBuilder;

pub use super::v0_1::NsTableValidationError;

pub use super::v0_1::NumNss;

pub use super::v0_1::NumTxs;

pub use super::v0_1::NumTxsRange;

pub use super::v0_1::NumTxsUnchecked;

pub use super::v0_1::Payload;

pub use super::v0_1::PayloadByteLen;

pub use super::v0_1::TimeBasedUpgrade;

pub use super::v0_1::Transaction;

pub use super::v0_1::TxIndex;

pub use super::v0_1::TxIter;

pub use super::v0_1::TxPayload;

pub use super::v0_1::TxPayloadRange;

pub use super::v0_1::TxProof;

pub use super::v0_1::TxTableEntries;

pub use super::v0_1::TxTableEntriesRange;

pub use super::v0_1::Upgrade;

pub use super::v0_1::UpgradeMode;

pub use super::v0_1::UpgradeType;

pub use super::v0_1::ViewBasedUpgrade;

pub use super::v0_1::BLOCK_MERKLE_TREE_HEIGHT;

pub use super::v0_1::FEE_MERKLE_TREE_HEIGHT;

pub use super::v0_1::NS_ID_BYTE_LEN;

pub use super::v0_1::NS_OFFSET_BYTE_LEN;

pub use super::v0_1::NUM_NSS_BYTE_LEN;

pub use super::v0_1::NUM_TXS_BYTE_LEN;

pub use super::v0_1::TX_OFFSET_BYTE_LEN;

Modules

auction 🔒

chain_config 🔒

fee_info 🔒

header 🔒

solver 🔒

Structs

BidTx

A transaction to bid for the sequencing rights of a namespace. It is the signed form of BidTxBody. Expected usage is build it by calling signed on BidTxBody.

BidTxBody

A transaction body holding data required for bid submission.

ChainConfig

Global variables for an Espresso blockchain.

Header

A header is like a [Block] with the body replaced by a digest.

ResolvableChainConfig

A commitment to a ChainConfig or a full ChainConfig.

RollupRegistration

RollupRegistrationBody

RollupUpdate

RollupUpdatebody

SolverAuctionResults

The results of an Auction

Enums

FullNetworkTx

Wrapper enum for Full Network Transactions. Each transaction type will be a variant of this enum.

Constants

VERSION

Traits

IterableFeeInfo

Methods for use w/ Vec