Module espresso_types::v0

Re-exports

pub use v0_3::AccountQueryData;

pub use v0_3::BlockMerkleCommitment;

pub use v0_3::BlockMerkleTree;

pub use v0_3::BuilderSignature;

pub use v0_3::ChainId;

pub use v0_3::Delta;

pub use v0_3::FeeAccount;

pub use v0_3::FeeAccountProof;

pub use v0_3::FeeAmount;

pub use v0_3::FeeInfo;

pub use v0_3::FeeMerkleCommitment;

pub use v0_3::FeeMerkleProof;

pub use v0_3::FeeMerkleTree;

pub use v0_3::Index;

pub use v0_3::Iter;

pub use v0_3::L1BlockInfo;

pub use v0_3::L1Client;

pub use v0_3::L1Snapshot;

pub use v0_3::NamespaceId;

pub use v0_3::NsIndex;

pub use v0_3::NsIter;

pub use v0_3::NsPayload;

pub use v0_3::NsPayloadBuilder;

pub use v0_3::NsPayloadByteLen;

pub use v0_3::NsPayloadOwned;

pub use v0_3::NsPayloadRange;

pub use v0_3::NsProof;

pub use v0_3::NsTable;

pub use v0_3::NsTableBuilder;

pub use v0_3::NsTableValidationError;

pub use v0_3::NumNss;

pub use v0_3::NumTxs;

pub use v0_3::NumTxsRange;

pub use v0_3::NumTxsUnchecked;

pub use v0_3::Payload;

pub use v0_3::PayloadByteLen;

pub use v0_3::Transaction;

pub use v0_3::TxIndex;

pub use v0_3::TxIter;

pub use v0_3::TxPayload;

pub use v0_3::TxPayloadRange;

pub use v0_3::TxProof;

pub use v0_3::TxTableEntries;

pub use v0_3::TxTableEntriesRange;

pub use v0_3::Upgrade;

pub use v0_3::UpgradeType;

pub use v0_3::UpgradeMode;

pub use v0_3::TimeBasedUpgrade;

pub use v0_3::ViewBasedUpgrade;

pub use v0_3::BlockSize;

pub use crate::v0_1::BLOCK_MERKLE_TREE_HEIGHT;

pub use crate::v0_1::FEE_MERKLE_TREE_HEIGHT;

pub use crate::v0_1::NS_ID_BYTE_LEN;

pub use crate::v0_1::NS_OFFSET_BYTE_LEN;

pub use crate::v0_1::NUM_NSS_BYTE_LEN;

pub use crate::v0_1::NUM_TXS_BYTE_LEN;

pub use crate::v0_1::TX_OFFSET_BYTE_LEN;

Modules

header 🔒

impls 🔒

mock

traits

This module contains all the traits used for building the sequencer types. It also includes some trait implementations that cannot be implemented in an external crate.

utils 🔒

v0_1

v0_2

v0_3

Macros

assert_eq_all_versions_of_type 🔒

define_modules 🔒

reexport_latest_version_of_type 🔒

reexport_unchanged_types 🔒

Re-export types which have not changed across any minor version.

with_minor_versions 🔒

Structs

BackoffParams

GenesisHeader

Information about the genesis state which feeds into the genesis block header.

NodeState

Represents the immutable state of a node.

ParseDurationError

ParseSizeError

Ratio

SeqTypes

SequencerVersions

SolverAuctionResultsProvider

Auction Results provider holding the Url of the solver in order to fetch auction results.

Timestamp

ValidatedState

Enums

BuilderValidationError

Possible builder validation failures

FeeError

Possible charge fee failures

Header

Each variant represents a specific minor version header.

ParseRatioError

ProposalValidationError

Possible proposal validation failures

StateValidationError

This enum is not used in code but functions as an index of possible validation errors.

Update

Constants

BACKOFF_FACTOR

BACKOFF_JITTER

MAX_RETRY_DELAY

MIN_RETRY_DELAY

Functions

parse_duration

parse_size

validate_proposal

Type Aliases

Event

FeeVersion

Leaf

MarketplaceVersion

MockSequencerVersions

NetworkConfig

PrivKey

PubKey

V0_1