espresso_types::v0::impls

Module state

Structs

Proposal 🔒

Block Proposal to be verified and applied.

ValidatedState

State to be validated by replicas.

ValidatedTransition 🔒

Type to hold cloned validated state and provide validation methods.

Enums

BuilderValidationError

Possible builder validation failures

ProposalValidationError

Possible proposal validation failures

StateValidationError

This enum is not used in code but functions as an index of possible validation errors.

Functions

_apply_full_transactions 🔒

get_l1_deposits

validate_builder_fee 🔒

Validate builder accounts by verifying signatures. All fees are verified against signature by index.