espresso_types::v0::traits
Trait FromNsPayloadBytes

pub trait FromNsPayloadBytes<'a> {
    // Required method
    fn from_payload_bytes(bytes: &'a [u8]) -> Self;
}

Data that can be deserialized from a subslice of namespace payload bytes.

Companion trait for NsPayloadBytesRange, which specifies the subslice of namespace payload bytes to read.

Required Methods

fn from_payload_bytes(bytes: &'a [u8]) -> Self

Deserialize Self from namespace payload bytes.

Object Safety

This trait is not object safe.