Trait espresso_types::v0::traits::SequencerPersistence

pub trait SequencerPersistence: Sized + Send + Sync + 'static {

    // Required methods
    fn load_config<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<NetworkConfig>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn save_config<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        cfg: &'life1 NetworkConfig,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn collect_garbage<'life0, 'async_trait>(
        &'life0 mut self,
        view: ViewNumber,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn save_anchor_leaf<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        leaf: &'life1 Leaf,
        qc: &'life2 QuorumCertificate<SeqTypes>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn load_latest_acted_view<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<ViewNumber>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_anchor_leaf<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<(Leaf, QuorumCertificate<SeqTypes>)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_undecided_state<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<(CommitmentMap<Leaf>, BTreeMap<ViewNumber, View<SeqTypes>>)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_quorum_proposals<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<BTreeMap<ViewNumber, Proposal<SeqTypes, QuorumProposal<SeqTypes>>>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_vid_share<'life0, 'async_trait>(
        &'life0 self,
        view: ViewNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Proposal<SeqTypes, VidDisperseShare<SeqTypes>>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn load_da_proposal<'life0, 'async_trait>(
        &'life0 self,
        view: ViewNumber,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Proposal<SeqTypes, DaProposal<SeqTypes>>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn append_vid<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        proposal: &'life1 Proposal<SeqTypes, VidDisperseShare<SeqTypes>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn append_da<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        proposal: &'life1 Proposal<SeqTypes, DaProposal<SeqTypes>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn record_action<'life0, 'async_trait>(
        &'life0 mut self,
        view: ViewNumber,
        action: HotShotAction,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_undecided_state<'life0, 'async_trait>(
        &'life0 mut self,
        leaves: CommitmentMap<Leaf>,
        state: BTreeMap<ViewNumber, View<SeqTypes>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn append_quorum_proposal<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        proposal: &'life1 Proposal<SeqTypes, QuorumProposal<SeqTypes>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn into_catchup_provider(
        self,
        _backoff: BackoffParams,
    ) -> Result<Arc<dyn StateCatchup>> { ... }
    fn load_consensus_state<'life0, 'async_trait>(
        &'life0 self,
        state: NodeState,
    ) -> Pin<Box<dyn Future<Output = Result<HotShotInitializer<SeqTypes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn handle_event<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: &'life1 Event,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }

}

Required Methods

fn load_config<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<NetworkConfig>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load the orchestrator config from storage.

Returns None if no config exists (we are joining a network for the first time). Fails with Err if it could not be determined whether a config exists or not.

fn save_config<'life0, 'life1, 'async_trait>( &'life0 mut self, cfg: &'life1 NetworkConfig, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Save the orchestrator config to storage.

fn collect_garbage<'life0, 'async_trait>( &'life0 mut self, view: ViewNumber, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn save_anchor_leaf<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, leaf: &'life1 Leaf, qc: &'life2 QuorumCertificate<SeqTypes>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Saves the latest decided leaf.

If the height of the new leaf is not greater than the height of the previous decided leaf, storage is not updated.

fn load_latest_acted_view<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<ViewNumber>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load the highest view saved with save_voted_view.

fn load_anchor_leaf<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<(Leaf, QuorumCertificate<SeqTypes>)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load the latest leaf saved with save_anchor_leaf.

fn load_undecided_state<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<(CommitmentMap<Leaf>, BTreeMap<ViewNumber, View<SeqTypes>>)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load undecided state saved by consensus before we shut down.

fn load_quorum_proposals<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Option<BTreeMap<ViewNumber, Proposal<SeqTypes, QuorumProposal<SeqTypes>>>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load the proposals saved by consensus

fn load_vid_share<'life0, 'async_trait>( &'life0 self, view: ViewNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<Proposal<SeqTypes, VidDisperseShare<SeqTypes>>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn load_da_proposal<'life0, 'async_trait>( &'life0 self, view: ViewNumber, ) -> Pin<Box<dyn Future<Output = Result<Option<Proposal<SeqTypes, DaProposal<SeqTypes>>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn append_vid<'life0, 'life1, 'async_trait>( &'life0 mut self, proposal: &'life1 Proposal<SeqTypes, VidDisperseShare<SeqTypes>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn append_da<'life0, 'life1, 'async_trait>( &'life0 mut self, proposal: &'life1 Proposal<SeqTypes, DaProposal<SeqTypes>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn record_action<'life0, 'async_trait>( &'life0 mut self, view: ViewNumber, action: HotShotAction, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn update_undecided_state<'life0, 'async_trait>( &'life0 mut self, leaves: CommitmentMap<Leaf>, state: BTreeMap<ViewNumber, View<SeqTypes>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn append_quorum_proposal<'life0, 'life1, 'async_trait>( &'life0 mut self, proposal: &'life1 Proposal<SeqTypes, QuorumProposal<SeqTypes>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Provided Methods

fn into_catchup_provider( self, _backoff: BackoffParams, ) -> Result<Arc<dyn StateCatchup>>

Use this storage as a state catchup backend, if supported.

fn load_consensus_state<'life0, 'async_trait>( &'life0 self, state: NodeState, ) -> Pin<Box<dyn Future<Output = Result<HotShotInitializer<SeqTypes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Load the latest known consensus state.

Returns an initializer to resume HotShot from the latest saved state (or start from genesis, if there is no saved state).

fn handle_event<'life0, 'life1, 'async_trait>( &'life0 mut self, event: &'life1 Event, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Update storage based on an event from consensus.

Object Safety

This trait is not object safe.