Trait espresso_types::v0::traits::StateCatchup

pub trait StateCatchup: Send + Sync + Debug {
    // Required methods
    fn try_fetch_account<'life0, 'async_trait>(
        &'life0 self,
        height: u64,
        view: ViewNumber,
        fee_merkle_tree_root: FeeMerkleCommitment,
        account: FeeAccount,
    ) -> Pin<Box<dyn Future<Output = Result<AccountQueryData>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn try_remember_blocks_merkle_tree<'life0, 'life1, 'async_trait>(
        &'life0 self,
        height: u64,
        view: ViewNumber,
        mt: &'life1 mut BlockMerkleTree,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn try_fetch_chain_config<'life0, 'async_trait>(
        &'life0 self,
        commitment: Commitment<ChainConfig>,
    ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn backoff(&self) -> &BackoffParams;

    // Provided methods
    fn fetch_accounts<'life0, 'async_trait>(
        &'life0 self,
        height: u64,
        view: ViewNumber,
        fee_merkle_tree_root: FeeMerkleCommitment,
        accounts: Vec<FeeAccount>,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountQueryData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn remember_blocks_merkle_tree<'life0, 'life1, 'async_trait>(
        &'life0 self,
        height: u64,
        view: ViewNumber,
        mt: &'life1 mut BlockMerkleTree,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn fetch_chain_config<'life0, 'async_trait>(
        &'life0 self,
        commitment: Commitment<ChainConfig>,
    ) -> Pin<Box<dyn Future<Output = ChainConfig> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

Required Methods

fn try_fetch_account<'life0, 'async_trait>( &'life0 self, height: u64, view: ViewNumber, fee_merkle_tree_root: FeeMerkleCommitment, account: FeeAccount, ) -> Pin<Box<dyn Future<Output = Result<AccountQueryData>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Try to fetch the given account state, failing without retrying if unable.

fn try_remember_blocks_merkle_tree<'life0, 'life1, 'async_trait>( &'life0 self, height: u64, view: ViewNumber, mt: &'life1 mut BlockMerkleTree, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Try to fetch and remember the blocks frontier, failing without retrying if unable.

fn try_fetch_chain_config<'life0, 'async_trait>( &'life0 self, commitment: Commitment<ChainConfig>, ) -> Pin<Box<dyn Future<Output = Result<ChainConfig>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn backoff(&self) -> &BackoffParams

Provided Methods

fn fetch_accounts<'life0, 'async_trait>( &'life0 self, height: u64, view: ViewNumber, fee_merkle_tree_root: FeeMerkleCommitment, accounts: Vec<FeeAccount>, ) -> Pin<Box<dyn Future<Output = Result<Vec<AccountQueryData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetch the given list of accounts, retrying on transient errors.

fn remember_blocks_merkle_tree<'life0, 'life1, 'async_trait>( &'life0 self, height: u64, view: ViewNumber, mt: &'life1 mut BlockMerkleTree, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetch and remember the blocks frontier, retrying on transient errors.

fn fetch_chain_config<'life0, 'async_trait>( &'life0 self, commitment: Commitment<ChainConfig>, ) -> Pin<Box<dyn Future<Output = ChainConfig> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,