Module espresso_types::v0::v0_1

Modules

block 🔒

chain_config 🔒

fee_info 🔒

header 🔒

instance_state 🔒

l1 🔒

signature 🔒

state 🔒

transaction 🔒

Structs

AccountQueryData

BlockSize

ChainConfig

Global variables for an Espresso blockchain.

ChainId

Delta

FeeAccount

FeeAccountProof

A proof of the balance of an account in the fee ledger.

FeeAmount

FeeInfo

FeeInfo holds data related to builder fees.

Header

A header is like a [Block] with the body replaced by a digest.

Index

Iter

Cartesian product of NsIter, TxIter.

L1BlockInfo

L1Client

An Http Provider and configuration to interact with the L1.

L1Snapshot

NamespaceId

NsIndex

Index for an entry in a ns table.

NsIter

Return type for [Payload::ns_iter].

NsPayload

Raw binary data for a single namespace’s payload.

NsPayloadBuilder

Build an individual namespace payload one transaction at a time.

NsPayloadByteLen

Byte length of a namespace payload.

NsPayloadOwned

NsPayloadRange

Index range for a namespace payload inside a block payload.

NsProof

Proof of correctness for namespace payload bytes in a block.

NsTable

Raw binary data for a namespace table.

NsTableBuilder

NumNss

Number of entries in a namespace table.

NumTxs

Number of txs in a namespace.

NumTxsRange

Byte range for the part of a tx table that declares the number of txs in the payload.

NumTxsUnchecked

The part of a tx table that declares the number of txs in the payload.

Payload

Raw payload data for an entire block.

PayloadByteLen

Byte length of a block payload, which includes all namespaces but not the namespace table.

ResolvableChainConfig

TimeBasedUpgrade

Represents an upgrade based on time (unix timestamp).

Transaction

TxIndex

Index for an entry in a tx table.

TxIter

TxPayload

A transaction’s payload data.

TxPayloadRange

Byte range for a transaction’s payload data.

TxProof

Proof of correctness for transaction bytes in a block.

TxTableEntries

Entries from a tx table in a namespace for use in a transaction proof.

TxTableEntriesRange

Byte range for entries from a tx table for use in a transaction proof.

Upgrade

Represents a general upgrade with mode and type.

ViewBasedUpgrade

Represents an upgrade based on view.

Enums

FeeMerkleProof

NsTableValidationError

Return type for NsTable::validate.

UpgradeMode

Represents the specific type of upgrade.

UpgradeType

Represents the specific type of upgrade.

Constants

BLOCK_MERKLE_TREE_HEIGHT

FEE_MERKLE_TREE_HEIGHT

NS_ID_BYTE_LEN

NS_OFFSET_BYTE_LEN

NUM_NSS_BYTE_LEN

Byte lengths for the different items that could appear in a namespace table.

NUM_TXS_BYTE_LEN

Byte lengths for the different items that could appear in a tx table.

TX_OFFSET_BYTE_LEN

VERSION

Type Aliases

BlockMerkleCommitment

BlockMerkleTree

BuilderSignature

FeeMerkleCommitment

FeeMerkleTree