Module espresso_types::v0::v0_1::block

Structs

Index

Iter

Cartesian product of NsIter, TxIter.

NsIndex

Index for an entry in a ns table.

NsIter

Return type for [Payload::ns_iter].

NsPayload

Raw binary data for a single namespace’s payload.

NsPayloadBuilder

Build an individual namespace payload one transaction at a time.

NsPayloadByteLen

Byte length of a namespace payload.

NsPayloadOwned

NsPayloadRange

Index range for a namespace payload inside a block payload.

NsProof

Proof of correctness for namespace payload bytes in a block.

NsTable

Raw binary data for a namespace table.

NsTableBuilder

NumNss

Number of entries in a namespace table.

NumTxs

Number of txs in a namespace.

NumTxsRange

Byte range for the part of a tx table that declares the number of txs in the payload.

NumTxsUnchecked

The part of a tx table that declares the number of txs in the payload.

Payload

Raw payload data for an entire block.

PayloadByteLen

Byte length of a block payload, which includes all namespaces but not the namespace table.

TxIndex

Index for an entry in a tx table.

TxIter

TxPayload

A transaction’s payload data.

TxPayloadRange

Byte range for a transaction’s payload data.

TxProof

Proof of correctness for transaction bytes in a block.

TxTableEntries

Entries from a tx table in a namespace for use in a transaction proof.

TxTableEntriesRange

Byte range for entries from a tx table for use in a transaction proof.

Enums

NsTableValidationError

Return type for NsTable::validate.

Constants

NS_ID_BYTE_LEN

NS_OFFSET_BYTE_LEN

NUM_NSS_BYTE_LEN

Byte lengths for the different items that could appear in a namespace table.

NUM_TXS_BYTE_LEN

Byte lengths for the different items that could appear in a tx table.

TX_OFFSET_BYTE_LEN