espresso_types::v0::v0_1
Module l1

Structs

L1BlockInfo

L1Client

An Ethereum provider and configuration to interact with the L1.

L1ClientMetrics 🔒

L1ClientOptions

Configuration for an L1 client.

L1Snapshot

L1State 🔒

In-memory view of the L1 state, updated asynchronously.

L1UpdateTask 🔒

SingleTransport 🔒

The state of the current provider being used by a SwitchingTransport. This is cloneable and returns a reference to the same underlying data.

SingleTransportStatus 🔒

The status of a single transport

SwitchingTransport

An RPC client with multiple remote (HTTP) providers.

Enums

L1Event 🔒